01_teams_bot/requirements.txt
botbuilder-core
botbuilder-schema
botbuilder-integration-flask
python-dotenv
google-generativeai
Flask
gunicorn